requirements.txt
pykrx>=1.0.45
yfinance>=1.2.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
feedparser>=6.0.0
beautifulsoup4>=4.12.0
google-genai>=1.0.0
python-telegram-bot>=21.0
schedule>=1.2.0
python-dotenv>=1.0.0
xgboost>=2.0.0
scikit-learn>=1.3.0
dart-fss>=0.4.3
anthropic>=0.40.0
fpdf2>=2.8.0
exchange_calendars>=4.5.0
fear_and_greed>=0.0.8
cot_reports>=0.3.2
